Update VMKeystrokes.ps1
Using Register-AutoCompleter, this will filter $hidCharacterMap to generate dynamic valid inputs and text completion for the SpecialKeyInput parameter. To accomplish this we need to move the $hidCharacterMap outside of the Set-VMKeystrokes function.

The regex filter selects items that are only letters and any values above 1 (removing all single character keys). Giving us a valid list of inputs based on the existing entries in $hidCharacterMap.